requirements.txt
Django==1.11
psycopg2==2.7.1
django-suit==0.2.25
Pillow==4.1.1
clarifai==2.0.21
django-graphos=0.3.41
